significant performance improvements:
```
benchcmp v1.txt clean-cache.txt
benchmark                                                 old ns/op     new ns/op     delta
BenchmarkAggregator1Aggregates2PointsPerAggregate-8       11284         10015         -11.25%
BenchmarkAggregator5Aggregates10PointsPerAggregate-8      8561049       135632        -98.42%
BenchmarkAggregator5Aggregates100PointsPerAggregate-8     15326290      1187623       -92.25%
```

when enabling the regex cache, duration is cut in half yet again (in exchange for more RAM usage)
```
cat clean-cache.txt
goos: linux
goarch: amd64
pkg: github.com/graphite-ng/carbon-relay-ng/aggregator
BenchmarkAggregator1Aggregates2PointsPerAggregate-8                	  200000	     10015 ns/op
BenchmarkAggregator5Aggregates10PointsPerAggregate-8               	   10000	    135632 ns/op
BenchmarkAggregator5Aggregates100PointsPerAggregate-8              	    1000	   1187623 ns/op
BenchmarkAggregator1Aggregates2PointsPerAggregateWithReCache-8     	  300000	      5761 ns/op
BenchmarkAggregator5Aggregates10PointsPerAggregateWithReCache-8    	   20000	     78453 ns/op
BenchmarkAggregator5Aggregates100PointsPerAggregateWithReCache-8   	    3000	    496872 ns/op
PASS
ok  	github.com/graphite-ng/carbon-relay-ng/aggregator	10.906s
```